add service getAllMessageOfRoom and getAllRoom

+++ BackEnd/Zalo App/src/main/java/com/essay/zaloapp/repository/CategoryUserRepository.java
@@ -3,5 +3,5 @@ package com.essay.zaloapp.repository;
 import com.essay.zaloapp.domain.models.CategoryUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryUserRepository extends JpaRepository<CategoryUser, Integer> {
+public interface CategoryUserRepository extends JpaRepository<CategoryUser, Long> {
 }
