+++ BackEnd/Zalo App/src/main/java/com/essay/zaloapp/repository/GroupChatRepository.java
@@ -3,5 +3,7 @@ package com.essay.zaloapp.repository;
 import com.essay.zaloapp.domain.models.GroupChat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupChatRepository extends JpaRepository<GroupChat, Integer> {
+import java.util.List;
+
+public interface GroupChatRepository extends JpaRepository<GroupChat, Long> {
 }
