+++ BackEnd/Zalo App/src/main/java/com/essay/zaloapp/repository/MessageTypeRepository.java
@@ -3,5 +3,5 @@ package com.essay.zaloapp.repository;
 import com.essay.zaloapp.domain.models.MessageType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MessageTypeRepository extends JpaRepository<MessageType, Integer> {
+public interface MessageTypeRepository extends JpaRepository<MessageType, Long> {
 }
